% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{find_taxon_rank_label}
\alias{find_taxon_rank_label}
\title{Find the taxonomic label given a Taxon ID and a Rank level}
\usage{
find_taxon_rank_label(ta, tid = NA, rank)
}
\arguments{
\item{ta}{a tidytacos object}

\item{tid}{Taxon ID}

\item{rank}{Taxonomic Rank}
}
\description{
Find the taxonomic label given a Taxon ID and a Rank level
}
