% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{my_add_top_genus}
\alias{my_add_top_genus}
\title{Find the two most abundant genus for each sample and add it to the samples table}
\usage{
my_add_top_genus(ta, genus_aggregation = F, dominance = T, dominance_thr = 0.3)
}
\arguments{
\item{ta}{a tidytacos object}

\item{genus_aggregation}{TRUE if table needs aggregation to the genus level.
FALSE if the provided tidytacos object is already aggregated.}

\item{dominance}{TRUE if a column should be added to specify the dominant taxon
per sample.}

\item{dominance_thr}{a numeric value specifying the minimum relative abundance
of the most abundant taxon to be considered as dominant.}
}
\description{
Find the two most abundant genus for each sample and add it to the samples table
}
