% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{plot_disc_col_from_df}
\alias{plot_disc_col_from_df}
\title{Generate a plot for a discrete column of a dataframe}
\usage{
plot_disc_col_from_df(my_df, my_col_ix)
}
\arguments{
\item{my_df}{a dataframe}

\item{my_col_ix}{a column index}
}
\description{
Generate a plot for a discrete column of a dataframe
}
