% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{plot_num_col_from_df}
\alias{plot_num_col_from_df}
\title{Generate a plot for a column of a dataframe}
\usage{
plot_num_col_from_df(
  my_df,
  my_col_ix,
  plot_type = c("histogram", "density", "boxplot"),
  n_bins = 30
)
}
\arguments{
\item{my_df}{a dataframe}

\item{my_col_ix}{a column index}

\item{plot_type}{the type of plot}

\item{n_bins}{the number of bins if a histogram is selected}
}
\description{
Generate a plot for a column of a dataframe
}
