% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{read_tidytacos_files}
\alias{read_tidytacos_files}
\title{Upload files from a folder created by tidytacos}
\usage{
read_tidytacos_files(samples, taxa, counts)
}
\arguments{
\item{samples}{the samples.csv file}

\item{taxa}{the taxa.csv file}

\item{counts}{the counts.csv file}
}
\description{
Upload files from a folder created by tidytacos
}
