% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdata.R
\docType{data}
\name{vdata}
\alias{vdata}
\title{vdata}
\format{
A \code{tidytacos} object with 3 data frames:
\describe{
\item{samples}{}
\item{taxa}{}
\item{counts}{}
}
}
\source{
\url{https://github.com/LebeerLab/Citizen-science-map-of-the-vaginal-microbiome}
}
\usage{
vdata
}
\description{
Amplicon count data of the paper "A citizen-science-enabled catalogue of the
vaginal microbiome and associated factors"
(\url{https://doi.org/10.1038/s41564-023-01500-0}). This dataset only
contains 200 samples.
}
\keyword{datasets}
